pyproject.toml
[project]
name = "crswrk-5"
version = "0.1.0"
description = ""
authors = [
    {name = "K",email = "kssshaab@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "django (>=4.2,<6.0)",
    "poetry-core (>=2.0.0,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "coverage (>=7.13.1,<8.0.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)",
    "redis (>=7.1.0,<8.0.0)",
    "celery (>=5.6.2,<6.0.0)",
    "django-celery-beat (>=2.8.1,<3.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "pillow (>=12.1.0,<13.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "types-requests (>=2.32.4.20260107,<3.0.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-django (>=4.12.0,<5.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
black = "^26.1.0"
flake8 = "^7.3.0"
isort = "^7.0.0"
mypy = "^1.19.1"

[tool.black]
line-length = 119
exclude = '''
/(
    \.venv
  | venv
  | \.git
  | \.mypy_cache
  | \.pytest_cache
  | build
  | dist
)/
'''

[tool.isort]
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv|README\.md'
ignore_missing_imports = true

